Refactor: move hankeData type definition under /domain/types;
This type is shared by components under /domain/
