+++ src/domain/hanke/edit/Form4.tsx
@@ -4,15 +4,14 @@ import { $enum } from 'ts-enum-util';
 import DatePicker from '../../../common/components/datePicker/DatePicker';
 import Dropdown from '../../../common/components/dropdown/Dropdown';
 import H2 from '../../../common/components/text/H2';
+import { FormProps, FORMFIELD } from './types';
 import {
-  FormProps,
-  FORMFIELD,
   HANKE_KAISTAHAITTA,
   HANKE_KAISTAPITUUSHAITTA,
   HANKE_MELUHAITTA,
   HANKE_POLYAITTA,
   HANKE_TARINAHAITTA,
-} from './types';
+} from '../../types/hanke';
 
 const Form4: React.FC<FormProps> = ({ control, errors, formData }) => {
   const { t, i18n } = useTranslation();
